[sync-api]add starcoin-sync-api.
