+++ chain/Cargo.toml
@@ -39,6 +39,7 @@ starcoin-statedb = {path = "../state/statedb"}
 bus = {path = "../bus", package="starcoin-bus" }
 starcoin-genesis = {path = "../core/genesis"}
 libra-temppath = { package="libra-temppath",  git = "https://github.com/starcoinorg/libra", rev="7a35ed6edb7921ab096028aed91b8e5ff11cc74d" }
+starcoin-sync-api = {package="starcoin-sync-api", path="../sync/api"}
 
 [dev-dependencies]
 starcoin-wallet-api = { path = "../wallet/api"}
