+++ sync/src/state_sync/mod.rs
@@ -3,13 +3,13 @@ use crate::{do_duration, DELAY_TIME};
 use actix::prelude::*;
 use actix::{Actor, Addr, Context, Handler};
 use anyhow::Result;
-use chain::SyncMetadata;
 use crypto::hash::HashValue;
 use executor::TransactionExecutor;
 use forkable_jellyfish_merkle::node_type::Node;
 use logger::prelude::*;
 use network::{NetworkAsyncService, RPCRequest, RPCResponse};
 use starcoin_state_tree::{StateNode, StateNodeStore};
+use starcoin_sync_api::SyncMetadata;
 use std::collections::VecDeque;
 use std::sync::Arc;
 use traits::Consensus;
